# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/traffic/countries": {
      "get": {
        "operationId": "TrafficController_getCountryTraffic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Total traffic distance per country",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryTrafficResponse"
                }
              }
            }
          }
        },
        "summary": "Total road traffic per country",
        "tags": [
          "traffic"
        ]
      }
    },
    "/traffic/vehicles": {
      "get": {
        "operationId": "TrafficController_getVehicleTraffic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Traffic distance per vehicle type, summed across countries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleTrafficResponse"
                }
              }
            }
          }
        },
        "summary": "Traffic distribution by vehicle type",
        "tags": [
          "traffic"
        ]
      }
    },
    "/traffic/observations/{countryCode}/{vehicleCode}/{year}": {
      "put": {
        "operationId": "TrafficController_updateObservation",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "vehicleCode",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "TOTAL",
                "LOR_RTRN",
                "MOTO_MOP",
                "CAR",
                "BUS_MCO_TRO"
              ],
              "type": "string"
            }
          },
          {
            "name": "countryCode",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "EE",
                "ES",
                "HU",
                "RO",
                "SE",
                "CH"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "number",
                    "minimum": 0
                  }
                },
                "required": [
                  "value"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The stored observation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Observation"
                }
              }
            }
          }
        },
        "summary": "Create or replace a single observation",
        "tags": [
          "traffic"
        ]
      }
    }
  },
  "info": {
    "title": "Derq Traffic API",
    "description": "Road traffic volumes from a fixed snapshot of Eurostat road_tf_vehmov.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "traffic",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CountryTrafficResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "countryCode": {
                  "type": "string"
                },
                "countryName": {
                  "type": "string"
                },
                "value": {
                  "type": "number"
                }
              },
              "required": [
                "countryCode",
                "countryName",
                "value"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "year": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "unitCode": {
                "type": "string"
              },
              "unitLabel": {
                "type": "string"
              }
            },
            "required": [
              "year",
              "unitCode",
              "unitLabel"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "VehicleTrafficResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "vehicleCode": {
                  "type": "string"
                },
                "vehicleName": {
                  "type": "string"
                },
                "value": {
                  "type": "number"
                }
              },
              "required": [
                "vehicleCode",
                "vehicleName",
                "value"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "year": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "unitCode": {
                "type": "string"
              },
              "unitLabel": {
                "type": "string"
              }
            },
            "required": [
              "year",
              "unitCode",
              "unitLabel"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "Observation": {
        "type": "object",
        "properties": {
          "countryCode": {
            "type": "string"
          },
          "vehicleCode": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "countryCode",
          "vehicleCode",
          "year",
          "value"
        ],
        "additionalProperties": false
      }
    }
  }
}